Combining the Typescript code down to one file.
